fix authorization of workProfile

+++ app/Http/Controllers/FreelancerController.php
@@ -10,6 +10,7 @@ use App\Models\Freelancer;
 use App\Models\Project;
 use App\Models\Room;
 use Carbon\Carbon;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
@@ -129,14 +130,29 @@ class FreelancerController extends Controller
                 'zip_code',
                 'location',
                 'note',
-                'salary_per_hour',
-                'salary_description',
-                'can_master'
             ]));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
+    public function updateTerms(Freelancer $freelancer, Request $request): void
+    {
+        $this->authorize('updateTerms', Freelancer::class);
+
+        $freelancer->update($request->only([
+            'salary_per_hour',
+            'salary_description',
+        ]));
+    }
+
+    /**
+     * @throws AuthorizationException
+     */
     public function updateWorkProfile(Freelancer $freelancer, Request $request): RedirectResponse
     {
+        $this->authorize('updateWorkProfile', Freelancer::class);
+
         $freelancer->update([
             'work_name' => $request->get('workName'),
             'work_description' => $request->get('workDescription')
@@ -145,8 +161,13 @@ class FreelancerController extends Controller
         return Redirect::back()->with('success', ['workProfile' => 'Arbeitsprofil erfolgreich aktualisiert']);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function updateCraftSettings(Freelancer $freelancer, Request $request): RedirectResponse
     {
+        $this->authorize('updateWorkProfile', Freelancer::class);
+
         $freelancer->update([
             'can_work_shifts' => $request->boolean('canBeAssignedToShifts'),
             'can_master' => $request->boolean('canBeUsedAsMasterCraftsman')
@@ -155,8 +176,13 @@ class FreelancerController extends Controller
         return Redirect::back();
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function assignCraft(Freelancer $freelancer, Request $request): RedirectResponse
     {
+        $this->authorize('updateWorkProfile', Freelancer::class);
+
         $craftToAssign = Craft::find($request->get('craftId'));
 
         if (is_null($craftToAssign)) {
@@ -170,8 +196,13 @@ class FreelancerController extends Controller
         return Redirect::back()->with('success', ['craft' => 'Gewerk erfolgreich zugeordnet.']);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function removeCraft(Freelancer $freelancer, Craft $craft): RedirectResponse
     {
+        $this->authorize('updateWorkProfile', Freelancer::class);
+
         $freelancer->assignedCrafts()->detach($craft);
 
         return Redirect::back()->with('success', ['craft' => 'Gewerk erfolgreich entfernt.']);
